first commit of a functional website with a backend
so you might read this commit message and think i have all of a sudden gotten somewhat good at this shit but not really. i just worked on with Navin on a mirror of a website that takes in username and password before accessing several pages. also implemented sessions so that i can keep contents of the websites very secure so that those jsps can't be accessed without logging in. the next thing i want to try now is to connect this with a database so that it can check if the username and password is there before you can be granted access. after that, i will work on a sign-in page and then your username and password gets stored there.