second commit.
i made a lot of changes and i wasn't committing as i was working which is really unfortunate but here i am to commit to you, my love. i am no longer afraid of commitments. i commit to only you, well not only you but other repos as well. i have to be honest with myself. well, the final change i made to this repo was working on a signup page. what happens when you don't have the credentials to log in to the website, you get redirected to a signup page and after putting in the username and password that you want, and then submitting, you then transferred back to the login page and then you input the details you just signed up with and you get access to the website. worked with databases here. i think i have a solid grasp on the concept of JDBC and DAO.