Able to confirm password in the sign up page.
it is not like i couldn't do this all this time but i just thought that creating a different jsp page that takes in the password you want to confirm against the password you inputted in the signUp.jsp wouldn't b such a bad idea to implement. had lots of problems with this and couldn't implement it in the end. i have had to settle for this which is funny because i saw this days ago when i wanted to sign up to a website.

i added a new file which is the signUpError.jsp just to alert the user of what was happening that the password was being an issue.

Now, i want to create a new branch for implementing what i have learnt as regards Hibernate but i should also merge these two branches before creating a new branch for hibernate because i think this is good enough to merge(to prod if you like lol).

Co-authored-by: shreax <efih.tesla@gmail.com>
lmao! that end is funny because that is me.