Integrated Hibernate with the Login.java file
i integrated hibernate with the Login.java file and removed the need for Jdbc. after learning hibernate, this was the next step, use what i have learnt with the small project of the website that asks for your login details and then it checks the database using hibernate rather than Jdbc. figuring this out was fun.