Integrated Hibernate with the signUpServlet.java file
i integrated hibernate with the signUpServlet.java file and removed the need for Jdbc. figuring this out was fun. i changed the data type of the id in the entity to be long rather than int because you never know the amount of users you might have signing up eventually.

so i came up with this idea of finding the number of rows in the database which is equal to the total number of users and using that figure(and adding one to it) to set the id of the new user.