fixed issue of what happens next if client's login details are incorrect
added just a few lines of code just to fix this issue. i was reading through the code when i noticed that there was an issue of what happens when the client's login details are incorrect. the page for loggin in is supposed to refresh at least. deployed the project just to confirm and the issue i suspected turned out to be actually there. i ended up fixing it by adding the else code which would redirect the client back to the login.jsp(the login page).